code merger into one file
